+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/webscanner.py
@@ -180,4 +180,4 @@ class WebScanner(Scanner):
     object_name_plural = pgettext_lazy("unit of scan", "web pages")
 
     class Meta:
-        verbose_name = _("Webscanner")
+        verbose_name = _("Web scanner")